Global explanations for tree-based models.

Global explanations for tree-based models by decomposing regression or classification functions into the sum of main components and interaction components of arbitrary order. Calculates SHAP values and q-interaction SHAP for all values of q for tree-based models such as xgboost.

Usage

glex(object, x, ...)

# S3 method for rpf
glex(object, x, max_interaction = NULL, ...)

# S3 method for xgb.Booster
glex(object, x, max_interaction = NULL, ...)

Arguments

object

Model to be explained.

x

Data to be explained.

...

Further arguments passed to methods.

max_interaction

(integer(1)) Maximum interaction size to consider. Defaults to using all possible interactions.

Value

Decomposition of the regression or classification function. A list with elements:

shap: SHAP values (xgboost method only).

m: Functional decomposition, i.e., all main and interaction components in the model, up to the degree specified by max_interaction.

intercept: Intercept, i.e., expected value of the prediction.

Details

For parallel execution, register a backend, e.g. with doParallel::registerDoParallel().

Examples

# Random Planted Forest -----
if (requireNamespace("randomPlantedForest", quietly = TRUE)) {
library(randomPlantedForest)

rp <- rpf(mpg ~ ., data = mtcars[1:26, ], max_interaction = 2)

glex_rpf <- glex(rp, mtcars[27:32, ])
str(glex_rpf, list.len = 5)
}
#> List of 3
#>  $ m        :Classes ‘data.table’ and 'data.frame':	6 obs. of  55 variables:
#>   ..$ cyl      : num [1:6] 0.884 0.884 -0.506 -0.297 -0.506 ...
#>   ..$ disp     : num [1:6] 0.632 2.864 -1.092 0.388 -0.911 ...
#>   ..$ hp       : num [1:6] 2.653 0.248 -2.573 -0.436 -2.573 ...
#>   ..$ drat     : num [1:6] 0.94 -0.034 0.94 0.1 0.1 ...
#>   ..$ wt       : num [1:6] 2.874 2.798 0.3 0.317 -0.882 ...
#>   .. [list output truncated]
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ intercept: num 20.2
#>  $ x        :Classes ‘data.table’ and 'data.frame':	6 obs. of  10 variables:
#>   ..$ cyl : num [1:6] 4 4 8 6 8 4
#>   ..$ disp: num [1:6] 120.3 95.1 351 145 301 ...
#>   ..$ hp  : num [1:6] 91 113 264 175 335 109
#>   ..$ drat: num [1:6] 4.43 3.77 4.22 3.62 3.54 4.11
#>   ..$ wt  : num [1:6] 2.14 1.51 3.17 2.77 3.57 ...
#>   .. [list output truncated]
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  - attr(*, "class")= chr [1:3] "glex" "rpf_components" "list"
# xgboost -----
if (requireNamespace("xgboost", quietly = TRUE)) {
library(xgboost)
x <- as.matrix(mtcars[, -1])
y <- mtcars$mpg
xg <- xgboost(data = x[1:26, ], label = y[1:26],
              params = list(max_depth = 4, eta = .1),
              nrounds = 10, verbose = 0)
glex(xg, x[27:32, ])

if (FALSE) {
# Parallel execution
doParallel::registerDoParallel()
glex(xg, x[27:32, ])
}
}