Print glex objects

This is implemented mainly to avoid flooding the console in cases where the glex object uses many terms, which leads to a large amount of column names of $m being printed to the console. This function wraps str() with a truncated output for a more compact representation.

Usage

# S3 method for glex
print(x, ...)

Arguments

x

Object to print.

...

(Unused)

Examples

# Random Planted Forest -----
if (requireNamespace("randomPlantedForest", quietly = TRUE)) {
library(randomPlantedForest)
rp <- rpf(mpg ~ hp + wt + drat, data = mtcars[1:26, ], max_interaction = 2)

glex(rp, mtcars[27:32, ])
}
#> glex object of subclass rpf_components 
#> Explaining predictions of 6 observations with 6 terms of up to 2 degrees
#> 
#> List of 3
#>  $ m        :Classes ‘data.table’ and 'data.frame':	6 obs. of  6 variables:
#>   ..$ hp     : num [1:6] 4.902 0.102 -3.918 -1.244 -3.918 ...
#>   ..$ wt     : num [1:6] 3.82 3.62 1.16 1.13 -1.7 ...
#>   ..$ drat   : num [1:6] 2.38 -0.45 2.38 -0.286 -0.286 ...
#>   ..$ hp:wt  : num [1:6] 0.386 -0.157 0.287 -0.309 -0.274 ...
#>   ..$ drat:hp: num [1:6] 0.6252 0.0806 0.0829 0.174 0.2327 ...
#>   .. [list output truncated]
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ intercept: num 20.1
#>  $ x        :Classes ‘data.table’ and 'data.frame':	6 obs. of  3 variables:
#>   ..$ hp  : num [1:6] 91 113 264 175 335 109
#>   ..$ wt  : num [1:6] 2.14 1.51 3.17 2.77 3.57 ...
#>   ..$ drat: num [1:6] 4.43 3.77 4.22 3.62 3.54 4.11
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  - attr(*, "class")= chr [1:3] "glex" "rpf_components" "list"